requirements.txt
#tensorflow-gpu
pandas
scikit-learn
seaborn
tensorflow-datasets
pydot
graphviz
pytz